+++ packages/core/styles/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
+## 2026-03-04 `9.61.0`
+
+#### 🎉 New features
+
+- `ProgressBar`: Added fill color CSS variable, background color variants, box shadow support, and no-overflow container style. ([#441](https://github.com/TiendaNube/nimbus-design-system/pull/441) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2026-03-02 `9.60.1`
 
 #### 🐛 Bug fixes